+++ BE/httc-sport/src/main/java/com/fpoly/httc_sport/repository/ChatRoomRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ChatRoomRepository extends JpaRepository<ChatRoom, Integer> {
+public interface ChatRoomRepository extends JpaRepository<ChatRoom, Long> {
 	List<ChatRoom> findByUserIdAndAdminId(String userId, String adminId);
 }
