+++ BE/httc-sport/src/main/java/com/fpoly/httc_sport/repository/PaymentMethodRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface PaymentMethodRepository extends JpaRepository<PaymentMethod, Integer> {
+public interface PaymentMethodRepository extends JpaRepository<PaymentMethod, Long> {
 	Optional<PaymentMethod> findByMethod(PaymentMethodEnum method);
 	boolean existsByMethod(PaymentMethodEnum method);
 }
